MRSW implemented

+++ 15/3/FileUploader/src/main/java/model/UploadModelImpl.java
@@ -1,7 +1,6 @@
 package model;
 
 import javax.annotation.ManagedBean;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -10,19 +9,39 @@ import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.locks.ReadWriteLock;
+import java.util.concurrent.locks.ReentrantReadWriteLock;
 
 @ManagedBean
 public class UploadModelImpl implements UploadModel {
-  private static volatile ConcurrentHashMap<String, Path> storage = new ConcurrentHashMap<>();
+  private static volatile ConcurrentHashMap<String, PathWithLock> storage = new ConcurrentHashMap<>();
 
   @Override
   public void writeFile(String fileName, InputStream inputStream) throws IOException {
-    Path pathToFile = Files.createTempFile(fileName, "tmp");
-    pathToFile.toFile().deleteOnExit();
+
+    final Path pathToFile;
+    final ReadWriteLock readWriteLock;
+    final PathWithLock pathWithLock = storage.get(fileName);
+
+    if (pathWithLock != null) {
+      pathToFile = pathWithLock.getPath();
+      readWriteLock = pathWithLock.getReadWriteLock();
+    } else {
+      pathToFile = Files.createTempFile(fileName, "tmp");
+      pathToFile.toFile().deleteOnExit();
+      readWriteLock = new ReentrantReadWriteLock();
+    }
+
+    readWriteLock.writeLock().lock();
+    storage.put(fileName, new PathWithLock(pathToFile, readWriteLock));
 
     try (OutputStream outputStream = Files.newOutputStream(pathToFile)) {
       inputStream.transferTo(outputStream);
-      storage.put(fileName, pathToFile);
+    } catch (IOException e) {
+      storage.remove(fileName);
+      throw e;
+    } finally {
+      readWriteLock.writeLock().unlock();
     }
   }
 
@@ -34,8 +53,15 @@ public class UploadModelImpl implements UploadModel {
   @Override
   public void loadFile(String fileName, OutputStream outputStream) throws IOException {
 
-    try (InputStream inputStream = Files.newInputStream(storage.get(fileName))) {
-      inputStream.transferTo(outputStream);
+    final PathWithLock pathWithLock = storage.get(fileName);
+
+    if (pathWithLock != null) {
+      pathWithLock.getReadWriteLock().readLock().lock();
+      try (InputStream inputStream = Files.newInputStream(pathWithLock.getPath())) {
+        inputStream.transferTo(outputStream);
+      } finally {
+        pathWithLock.getReadWriteLock().readLock().unlock();
+      }
     }
   }
